Bug 1783761 - extract out per_file_info.rs from output-file.rs
For directory consumption, we're going to want to have a more formal
data structure hierarchy for per-file info.  The plan is to then
serialize the aggregated information into the "description" files which
will become JSON.

The derivation of this information potentially might end up moving
into the crossref phase in the future.
