+++ tools/src/file_format/mod.rs
@@ -2,3 +2,4 @@ pub mod analysis;
 pub mod crossref_lookup;
 pub mod identifiers;
 pub mod merger;
+pub mod per_file_info;
